# SimBlog17API/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SimBlog17API",
    "version": "1.0.0",
    "description": "Lightweight forum backend API.\nSupports JWT access/refresh tokens, articles, comments, interactions, homepage aggregation, and admin moderation.\n"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
...
  "paths": {
...
    "/api/comments": {
      "post": {
        "tags": [
          "Comment"
        ],
        "summary": "Create comment or reply",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseCommentResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        }
      }
    },
    "/api/comments/by_article/{articleId}": {
      "get": {
        "tags": [
          "Comment"
        ],
        "summary": "List comments by article",
        "parameters": [
          {
            "in": "path",
            "name": "articleId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          },
          {
            "$ref": "#/components/parameters/Sort"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponsePageCommentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/comments/{id}": {
      "delete": {
        "tags": [
          "Comment"
        ],
        "summary": "Delete comment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseCommentResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/articles/{id}/like": {
      "post": {
        "tags": [
          "Interaction"
        ],
        "summary": "Toggle like",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ArticleIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseBoolean"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/articles/{id}/favorite": {
      "post": {
        "tags": [
          "Interaction"
        ],
        "summary": "Toggle favorite",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ArticleIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseBoolean"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/articles/{id}/interactions": {
      "get": {
        "tags": [
          "Interaction"
        ],
        "summary": "Get interaction status",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArticleIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseArticleInteractionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/articles/profile/{uid}/likes": {
      "get": {
        "tags": [
          "Interaction"
        ],
        "summary": "List user's liked articles",
        "parameters": [
          {
            "$ref": "#/components/parameters/UidPath"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          },
          {
            "$ref": "#/components/parameters/Sort"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponsePageArticleMetaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/articles/profile/{uid}/favorites": {
      "get": {
        "tags": [
          "Interaction"
        ],
        "summary": "List user's favorited articles",
        "parameters": [
          {
            "$ref": "#/components/parameters/UidPath"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          },
          {
            "$ref": "#/components/parameters/Sort"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponsePageArticleMetaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/home/latest": {
      "get": {
        "tags": [
          "Home"
        ],
        "summary": "Latest published articles",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          },
          {
            "$ref": "#/components/parameters/Sort"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponsePageArticleMetaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/home/hot": {
      "get": {
        "tags": [
          "Home"
        ],
        "summary": "Hot articles in N days",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 7,
              "minimum": 1,
              "maximum": 365
            }
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          },
          {
            "$ref": "#/components/parameters/Sort"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponsePageArticleMetaResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/home/hot-tags": {
      "get": {
        "tags": [
          "Home"
        ],
        "summary": "Hot tags",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseHotTagList"
                }
              }
            }
          }
        }
      }
    },
    "/api/home/recent-comments": {
      "get": {
        "tags": [
          "Home"
        ],
        "summary": "Recent comments",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          },
          {
            "$ref": "#/components/parameters/Sort"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponsePageCommentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/comments/{id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete comment as admin",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseCommentResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "InvalidRequest": {
        "description": "Invalid Request - validation or payload format error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiStatusResponseString"
            },
            "example": {
              "statusCode": 2,
              "statusMessage": "Invalid Request",
              "timeStamp": 1778003000000,
              "traceId": "f84f2d4d-7f13-4bd1-8b63-ec2f58b1f54b",
              "data": "Validation failed: username must not be blank"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - missing/invalid credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiStatusBase"
            },
            "example": {
              "statusCode": 5,
              "statusMessage": "Unauthorized",
              "timeStamp": 1778002000000,
              "traceId": "5f8a6ca2-b8bf-4ad4-b3d4-0d7f7dd6a20b"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiStatusBase"
            },
            "example": {
              "statusCode": 4,
              "statusMessage": "Access Denied",
              "timeStamp": 1778002000000,
              "traceId": "2dbf3d1d-06ee-4a8d-a865-1e8cc7f43a6d"
            }
          }
        }
      },
...
    },
    "parameters": {
      "UidPath": {
        "in": "path",
        "name": "uid",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
      "ArticleIdPath": {
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
      "Page": {
        "in": "query",
        "name": "page",
        "schema": {
          "type": "integer",
          "default": 0,
          "minimum": 0
        }
      },
      "Size": {
        "in": "query",
        "name": "size",
        "schema": {
          "type": "integer",
          "default": 10,
          "minimum": 1,
          "maximum": 200
        }
      },
      "Sort": {
        "in": "query",
        "name": "sort",
        "schema": {
          "type": "string",
          "example": "publishedTime,desc"
        }
      },
...
    },
    "schemas": {
...
      "ArticleContentType": {
        "type": "string",
        "enum": [
          "PLAIN_TEXT",
          "MARKDOWN",
          "HTML"
        ]
      },
...
      "ArticleMetaResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "preview": {
            "type": "string"
          },
          "contentType": {
            "$ref": "#/components/schemas/ArticleContentType"
          },
          "authorId": {
            "type": "integer",
            "format": "int64"
          },
          "publishedTime": {
            "type": "string",
            "format": "date-time"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          },
          "viewCount": {
            "type": "integer"
          },
          "tags": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          }
        }
      },
      "CommentRequest": {
        "type": "object",
        "required": [
          "articleId",
          "content"
        ],
        "properties": {
          "articleId": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string",
            "maxLength": 2000
          },
          "parentCommentId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        }
      },
      "CommentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "articleId": {
            "type": "integer",
            "format": "int64"
          },
          "authorId": {
            "type": "integer",
            "format": "int64"
          },
          "parentCommentId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "content": {
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer"
          }
        }
      },
...
      "ArticleInteractionResponse": {
        "type": "object",
        "properties": {
          "likeCount": {
            "type": "integer",
            "format": "int64"
          },
          "favoriteCount": {
            "type": "integer",
            "format": "int64"
          },
          "likedByCurrentUser": {
            "type": "boolean"
          },
          "favoritedByCurrentUser": {
            "type": "boolean"
          }
        }
      },
...
      "HomeHotTagEntry": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PageArticleMetaResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleMetaResponse"
            }
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer"
          },
          "number": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          }
        }
      },
...
      "PageCommentResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentResponse"
            }
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer"
          },
          "number": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          }
        }
      },
...
      "ApiStatusBase": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "statusMessage": {
            "type": "string"
          },
          "timeStamp": {
            "type": "integer",
            "format": "int64"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "ApiStatusResponseBoolean": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "boolean"
              }
            }
          }
        ]
      },
      "ApiStatusResponseString": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "ApiStatusResponseCommentResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/CommentResponse"
              }
            }
          }
        ]
      },
...
      "ApiStatusResponseArticleInteractionResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ArticleInteractionResponse"
              }
            }
          }
        ]
      },
...
      "ApiStatusResponseHotTagList": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/HomeHotTagEntry"
                }
              }
            }
          }
        ]
      },
      "ApiStatusResponsePageArticleMetaResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/PageArticleMetaResponse"
              }
            }
          }
        ]
      },
...
      "ApiStatusResponsePageCommentResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/PageCommentResponse"
              }
            }
          }
        ]
      },
...
    }
  }
}